Split LevelManager from RenderCore, both now having an initialize(). MarbleLevel now loads with default properties (fixes skylight before level loads on client). Added to initialization for client (incomplete) and editor.
