+++ src/client/editor/models.js
@@ -2,7 +2,7 @@ import * as THREE from "three";
 import "three/examples/js/loaders/GLTFLoader";
 import { editorLog } from "./log";
 import { projectTab } from "./project";
-import { renderCore } from "../render/render-core";
+import { levelManager } from "../level-manager";
 import { materialsTab } from "./materials";
 import "three/examples/js/QuickHull";
 import "three/examples/js/geometries/ConvexGeometry";
@@ -317,7 +317,7 @@ let modelsTab = function() {
 		initialize: function() {
 			_GLTFLoader = new THREE.GLTFLoader();
 			this.group = new THREE.Group();
-			renderCore.activeLevel.scene.add(this.group);
+			levelManager.activeLevel.scene.add(this.group);
 			this.group.visible = false;
 
 			// Add models button
