+++ src/client/editor/prefabs.js
@@ -4,6 +4,7 @@ import { hslToHex } from "../hsl-to-hex";
 import { inspector } from "./inspector";
 import * as materials from "./editor-materials";
 import { renderCore } from "../render/render-core";
+import { levelManager } from "../level-manager";
 import { EditorObject } from "./editor";
 import { modelsTab } from "./models";
 import { worldTab } from "./world";
@@ -573,7 +574,7 @@ let prefabsTab = function() {
 			this.elements.modelList = document.getElementById("inspectorModelList");
 			this.elements.colliderList = document.getElementById("inspectorColliderList");
 			this.group = new THREE.Group();
-			renderCore.activeLevel.scene.add(this.group);
+			levelManager.activeLevel.scene.add(this.group);
 			this.group.visible = false;
 
 			// Register new prefab event
